queuing strategy comments
