+++ modules/llrt_stream_web/src/queuing_strategy/mod.rs
@@ -1,11 +1,19 @@
 use rquickjs::{
     class::{JsCell, JsClass, Readable, Trace},
     function::{Constructor, Params},
-    methods, Class, Ctx, Error, Exception, FromJs, Function, JsLifetime, Object, Result, Value,
+    Class, Ctx, Error, Exception, FromJs, Function, JsLifetime, Object, Result, Value,
 };
 
 use super::ValueOrUndefined;
 
+pub(crate) use byte_length::ByteLengthQueuingStrategy;
+pub(crate) use count::CountQueuingStrategy;
+
+mod byte_length;
+mod count;
+
+/// QueuingStrategy is the structure of a user-provided object describing how backpressure should be signalled.
+/// https://streams.spec.whatwg.org/#qs-api
 pub(super) struct QueuingStrategy<'js> {
     // unrestricted double highWaterMark;
     high_water_mark: Option<Value<'js>>,
@@ -18,7 +26,7 @@ impl<'js> FromJs<'js> for QueuingStrategy<'js> {
         let ty_name = value.type_name();
         let obj = value
             .as_object()
-            .ok_or(Error::new_from_js(ty_name, "Object"))?;
+            .ok_or_else(|| Error::new_from_js(ty_name, "Object"))?;
 
         let high_water_mark = obj.get_value_or_undefined::<_, Value>("highWaterMark")?;
         let size = obj.get_value_or_undefined::<_, _>("size")?;
@@ -34,35 +42,34 @@ impl<'js> QueuingStrategy<'js> {
     // https://streams.spec.whatwg.org/#validate-and-normalize-high-water-mark
     pub(super) fn extract_high_water_mark(
         ctx: &Ctx<'js>,
-        this: Option<QueuingStrategy<'js>>,
+        this: Option<Self>,
         default_hwm: f64,
     ) -> Result<f64> {
         match this {
             // If strategy["highWaterMark"] does not exist, return defaultHWM.
             None => Ok(default_hwm),
-            Some(this) => {
-                // Let highWaterMark be strategy["highWaterMark"].
-                if let Some(high_water_mark) = &this.high_water_mark {
-                    let high_water_mark = high_water_mark.as_number().unwrap_or(f64::NAN);
-                    // If highWaterMark is NaN or highWaterMark < 0, throw a RangeError exception.
-                    if high_water_mark.is_nan() || high_water_mark < 0.0 {
-                        Err(Exception::throw_range(ctx, "Invalid highWaterMark"))
-                    } else {
-                        // Return highWaterMark.
-                        Ok(high_water_mark)
-                    }
+            // Let highWaterMark be strategy["highWaterMark"].
+            Some(Self {
+                high_water_mark: Some(high_water_mark),
+                ..
+            }) => {
+                let high_water_mark = high_water_mark.as_number().unwrap_or(f64::NAN);
+                // If highWaterMark is NaN or highWaterMark < 0, throw a RangeError exception.
+                if high_water_mark.is_nan() || high_water_mark < 0.0 {
+                    Err(Exception::throw_range(ctx, "Invalid highWaterMark"))
                 } else {
-                    // If strategy["highWaterMark"] does not exist, return defaultHWM.
-                    Ok(default_hwm)
+                    // Return highWaterMark.
+                    Ok(high_water_mark)
                 }
             },
+
+            // If strategy["highWaterMark"] does not exist, return defaultHWM.
+            _ => Ok(default_hwm),
         }
     }
 
     // https://streams.spec.whatwg.org/#make-size-algorithm-from-size-function
-    pub(super) fn extract_size_algorithm(
-        this: Option<&QueuingStrategy<'js>>,
-    ) -> SizeAlgorithm<'js> {
+    pub(super) fn extract_size_algorithm(this: Option<&Self>) -> SizeAlgorithm<'js> {
         // If strategy["size"] does not exist, return an algorithm that returns 1.
         match this.as_ref().and_then(|t| t.size.as_ref()) {
             None => SizeAlgorithm::AlwaysOne,
@@ -71,6 +78,7 @@ impl<'js> QueuingStrategy<'js> {
     }
 }
 
+/// SizeAlgorithm represents the two ways we might generate sizes - by calling a function or by simply returning 1.0 (the default)
 #[derive(JsLifetime, Trace, Clone)]
 pub(super) enum SizeAlgorithm<'js> {
     AlwaysOne,
@@ -93,6 +101,8 @@ impl<'js> SizeAlgorithm<'js> {
     }
 }
 
+/// SizeValue abstracts over the sources of size values - they can either come from user-provided size functions, in which case they might
+/// be any Value, or (more often) they come from a NativeSizeFunction or the default AlwaysOne algorithm and we can pass around a native Rust type.
 pub(super) enum SizeValue<'js> {
     Value(Value<'js>),
     Native(f64),
@@ -117,6 +127,8 @@ impl<'js> FromJs<'js> for SizeValue<'js> {
     }
 }
 
+/// SizeFunction abstracts over user-provided size functions (from their own queuing strategy implementations) and ones provided by us.
+/// We want to be able to recognise the ones that we have provided so we can short-circuit expensive JS calls
 #[derive(JsLifetime, Trace, Clone)]
 pub(super) enum SizeFunction<'js> {
     Js(Function<'js>),
@@ -133,36 +145,9 @@ impl<'js> FromJs<'js> for SizeFunction<'js> {
     }
 }
 
-#[derive(JsLifetime, Trace)]
-#[rquickjs::class]
-pub(crate) struct CountQueuingStrategy<'js> {
-    high_water_mark: f64,
-    size: Class<'js, NativeSizeFunction>,
-}
-
-#[methods(rename_all = "camelCase")]
-impl<'js> CountQueuingStrategy<'js> {
-    #[qjs(constructor)]
-    fn new(ctx: Ctx<'js>, init: QueueingStrategyInit) -> Result<Self> {
-        // Set this.[[highWaterMark]] to init["highWaterMark"].
-        Ok(Self {
-            high_water_mark: init.high_water_mark,
-            size: Class::instance(ctx, NativeSizeFunction::Count)?,
-        })
-    }
-
-    #[qjs(get)]
-    fn size(&self) -> Class<'js, NativeSizeFunction> {
-        self.size.clone()
-    }
-
-    #[qjs(get)]
-    fn high_water_mark(&self) -> f64 {
-        self.high_water_mark
-    }
-}
-
-struct QueueingStrategyInit {
+/// QueueingStrategyInit is the dictionary of input parameters for both native queuing strategies
+/// https://streams.spec.whatwg.org/#dictdef-queuingstrategyinit
+pub(crate) struct QueueingStrategyInit {
     high_water_mark: f64,
 }
 
@@ -171,45 +156,18 @@ impl<'js> FromJs<'js> for QueueingStrategyInit {
         let ty_name = value.type_name();
         let obj = value
             .as_object()
-            .ok_or(Error::new_from_js(ty_name, "Object"))?;
+            .ok_or_else(|| Error::new_from_js(ty_name, "Object"))?;
 
         let high_water_mark = obj
             .get_value_or_undefined("highWaterMark")?
-            .ok_or(Error::new_from_js(ty_name, "QueueingStrategyInit"))?;
+            .ok_or_else(|| Error::new_from_js(ty_name, "QueueingStrategyInit"))?;
 
         Ok(Self { high_water_mark })
     }
 }
 
-#[derive(JsLifetime, Trace)]
-#[rquickjs::class]
-pub(crate) struct ByteLengthQueuingStrategy<'js> {
-    high_water_mark: f64,
-    size: Class<'js, NativeSizeFunction>,
-}
-
-#[methods(rename_all = "camelCase")]
-impl<'js> ByteLengthQueuingStrategy<'js> {
-    #[qjs(constructor)]
-    fn new(ctx: Ctx<'js>, init: QueueingStrategyInit) -> Result<Self> {
-        // Set this.[[highWaterMark]] to init["highWaterMark"].
-        Ok(Self {
-            high_water_mark: init.high_water_mark,
-            size: Class::instance(ctx, NativeSizeFunction::ByteLength)?,
-        })
-    }
-
-    #[qjs(get)]
-    fn size(&self) -> Class<'js, NativeSizeFunction> {
-        self.size.clone()
-    }
-
-    #[qjs(get)]
-    fn high_water_mark(&self) -> f64 {
-        self.high_water_mark
-    }
-}
-
+/// NativeSizeFunction is a callable class which allows us to keep track that these size functions are not user provided, but
+/// in fact represent the native size functions. This allows us to avoid JS calls by noticing that a size function is this class.
 #[derive(JsLifetime, Trace, Clone, Copy)]
 pub(super) enum NativeSizeFunction {
     ByteLength,
